# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pennant",
    "version": "1",
    "description": "The presentation contract between the Pennant server and the Mac app (D-056). Generated by `npm run contract:build` from `server/contract/`; do not edit by hand."
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
  "paths": {
...
    "/api/v2/claims/{key}": {
      "get": {
        "operationId": "getClaimTrail",
        "summary": "The evidence trail behind an item, built on demand (an MLB need's responses); the key is the item's `evidence`.",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "An item's `evidence` key.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimTrail"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/club/{teamId}": {
      "get": {
        "operationId": "getClubReport",
        "summary": "Another club's report under our scouting: its masthead, how it wins and loses, its roster map, its recent moves, injuries, record against us and next series with us.",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "required": true,
            "description": "A team id (or `automatic` for the club the app follows).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClubReport"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/departments/{org}/{dept}": {
      "get": {
        "operationId": "getDepartmentReport",
        "summary": "One department's full report: prepared by, key figures, to decide, watching, what changed, what we can't see.",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "required": true,
            "description": "A team id, or `automatic` for the club the app follows (the configured club, else the one the save's human runs).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dept",
            "in": "path",
            "required": true,
            "description": "The department's id (`majorLeague`, `farm`, ...).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentReport"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/desk/{org}": {
      "get": {
        "operationId": "getDesk",
        "summary": "The GM's desk: every item to decide with its status (open, reviewed, deferred, handled in OOTP), in the stated order, and the items set aside. A status records attention only.",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "required": true,
            "description": "A team id, or `automatic` for the club the app follows (the configured club, else the one the save's human runs).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeskView"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "setDeskStatus",
        "summary": "Marks an item open, reviewed, deferred until a game date, or handled in OOTP (and keeps a note). Answers with the status it replaced and the request that undoes it. Nothing is written to OOTP.",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "required": true,
            "description": "A team id, or `automatic` for the club the app follows (the configured club, else the one the save's human runs).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeskChange"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/following": {
      "get": {
        "operationId": "getFollowing",
        "summary": "The clubs and players the GM follows in this save (the watchlist copied in, with its notes), and the division rivals suggested.",
        "tags": [
          "v2"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Following"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "follow",
        "summary": "Follows a club or a player, or changes a follow's note; answers with the request that undoes it.",
        "tags": [
          "v2"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowChange"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "unfollow",
        "summary": "Stops following a club or a player; answers with the request that undoes it.",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "description": "`club` or `player`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "The team id or the player id.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowChange"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/front-office/{org}": {
      "get": {
        "operationId": "getFrontOffice",
        "summary": "The Morning Report: its masthead (record, place, runs, last five, next game, trade deadline), the lede, \"How we win and lose\", the roster map, the desk (every department's items to decide, in a stated order) and one card per department.",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "required": true,
            "description": "A team id, or `automatic` for the club the app follows (the configured club, else the one the save's human runs).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrontOfficeSummary"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/rating-history/choice": {
      "post": {
        "operationId": "answerRatingHistoryOffer",
        "summary": "The GM's answer: carry a history over into this save (`adopt`, a question or a listed history), keep them apart (`fresh`, a question only), or undo a carry-over (`undo`). An answer that no longer fits, or names another save, is refused (400).",
        "tags": [
          "v2"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingHistoryChoice"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingHistoryView"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/search": {
      "get": {
        "operationId": "search",
        "summary": "Players (ours and the league's), clubs and views matching every word typed, each with where it opens; followed first, then a stated order.",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "description": "What was typed.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchAnswer"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/theme-packs/{org}": {
      "get": {
        "operationId": "getThemeChoices",
        "summary": "The themes a club can wear: its own colours (from the save) and each installed theme pack that fits it, with every appearance resolved; the packs refused, and why.",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "required": true,
            "description": "A team id, or `automatic` for the club the app follows (the configured club, else the one the save's human runs).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemeChoices"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "chooseTheme",
        "summary": "Chooses the theme a club wears (`club-colors` for its own colours). A pack not installed, refused, or made for another club is refused (400).",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "required": true,
            "description": "A team id, or `automatic` for the club the app follows (the configured club, else the one the save's human runs).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThemeChoice"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThemeChoices"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/wire/{org}": {
      "get": {
        "operationId": "getWire",
        "summary": "Around the league: moves (in the transaction log's own words), trades, injuries, streaks, awards, standings and news, since the last export unless asked, in the stated order (followed first when asked, then newest, then kind).",
        "tags": [
          "v2"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "required": true,
            "description": "A team id, or `automatic` for the club the app follows (the configured club, else the one the save's human runs).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "description": "A game date (`2040-5-1`), or `season`; the last export's day when left out.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "club",
            "in": "query",
            "required": false,
            "description": "Only entries naming this club (a team id).",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "description": "`move`, `trade`, `injury`, `streak`, `award`, `standings` or `news`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "followed",
            "in": "query",
            "required": false,
            "description": "`first` puts followed clubs and players first; `only` keeps only them.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Wire"
                }
              }
            }
          },
          "400": {
            "description": "The request could not be done",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "The per-launch token the app hands the sidecar on stdin."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong, as a sentence."
          },
          "detail": {
            "type": "string",
            "description": "The raw message behind it, for the log and a help tag (a `/v2` route's failure); never the visible line."
          }
        },
        "required": [
          "error"
        ],
        "description": "What a route answers when it cannot do what was asked (a 400)."
      },
...
      "ClaimTrail": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "importStamp": {
            "type": [
              "string",
              "null"
            ]
          },
          "reportStamp": {
            "type": "string"
          },
          "title": {
            "$ref": "#/components/schemas/Cell"
          },
          "headline": {
            "$ref": "#/components/schemas/Claim"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrailSection"
            }
          }
        },
        "required": [
          "key",
          "importStamp",
          "reportStamp",
          "title",
          "headline",
          "sections"
        ],
        "description": "The evidence trail behind an item, fetched on demand (`GET /api/v2/claims/:key`): too slow to put in a summary (an MLB need's staff recommendation takes seconds on a real save), so the inspector asks for it when the GM opens the item."
      },
...
      "ClubReport": {
        "type": "object",
        "properties": {
          "teamId": {
            "$ref": "#/components/schemas/Integer"
          },
          "club": {
            "type": "string"
          },
          "abbreviation": {
            "type": [
              "string",
              "null"
            ]
          },
          "importStamp": {
            "type": [
              "string",
              "null"
            ]
          },
          "reportStamp": {
            "type": "string",
            "description": "The build it comes from; moves with every rebuild."
          },
          "asOf": {
            "$ref": "#/components/schemas/Cell"
          },
          "followed": {
            "type": "boolean",
            "description": "The GM follows the club."
          },
          "ours": {
            "type": "boolean",
            "description": "This is the club the app follows (its own report is the Morning Report)."
          },
          "teamSeason": {
            "$ref": "#/components/schemas/TeamSeason"
          },
          "lede": {
            "$ref": "#/components/schemas/Claim"
          },
          "clubProfile": {
            "$ref": "#/components/schemas/ClubProfile"
          },
          "rosterMap": {
            "$ref": "#/components/schemas/RosterMap"
          },
          "scouting": {
            "$ref": "#/components/schemas/Claim",
            "description": "Whose eyes the ratings are and how much of the club our scouts see, with what they can't see as its basis."
          },
          "headToHead": {
            "$ref": "#/components/schemas/Claim",
            "description": "Their record against us this season; null for our own club."
          },
          "nextSeries": {
            "$ref": "#/components/schemas/Claim",
            "description": "Their next series with us; null for our own club, or when none is scheduled (`nextSeriesNote` says so)."
          },
          "nextSeriesNote": {
            "$ref": "#/components/schemas/Cell"
          },
          "moves": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WireEntry"
            },
            "description": "Their recent moves: the wire filtered to the club, newest first."
          },
          "movesNote": {
            "$ref": "#/components/schemas/Cell"
          },
          "injuries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClubInjury"
            }
          },
          "injuriesNote": {
            "$ref": "#/components/schemas/Cell"
          },
          "openWire": {
            "$ref": "#/components/schemas/Target",
            "description": "Open the wire filtered to the club."
          }
        },
        "required": [
          "teamId",
          "club",
          "abbreviation",
          "importStamp",
          "reportStamp",
          "asOf",
          "followed",
          "ours",
          "scouting",
          "moves",
          "injuries",
          "openWire"
        ],
        "description": "Another club's report (`GET /api/v2/club/:teamId`): the same masthead, \"How they win and lose\" and roster map as ours (the same modules, D-059), under our organization's scouting, with their recent moves, their injuries, their record against us and their next series with us. Nothing ranks the club or reads it as a buyer or a seller (D-060)."
      },
...
      "DeskView": {
        "type": "object",
        "properties": {
          "orgId": {
            "$ref": "#/components/schemas/Integer"
          },
          "importStamp": {
            "type": [
              "string",
              "null"
            ]
          },
          "reportStamp": {
            "type": "string"
          },
          "deskStamp": {
            "type": "string",
            "description": "Moves whenever a status or a note changes, so a client knows its copy is current."
          },
          "desk": {
            "$ref": "#/components/schemas/Desk"
          }
        },
        "required": [
          "orgId",
          "importStamp",
          "reportStamp",
          "deskStamp",
          "desk"
        ],
        "description": "The desk on its own (`GET /api/v2/desk/:org`): the same desk the Morning Report shows, with every status."
      },
...
      "ThemeChoice": {
        "type": "object",
        "properties": {
          "packId": {
            "type": "string"
          }
        },
        "required": [
          "packId"
        ],
        "description": "What `POST /api/v2/theme-packs/:org` takes: the pack the club should wear (`club-colors` for its own colours)."
      },
...
    }
  }
}